requirements.txt
# Основные зависимости
aiosqlite>=0.19.0
PyMuPDF>=1.23.0
openai>=1.0.0
python-multipart>=0.0.6
langchain-gigachat>=0.1.0
uvicorn[standard]>=0.22.0
jinja2>=3.1.2
langchain>=0.0.340
hvac>=2.0.0
opensearch-py>=2.0.0
pyyaml>=6.0.1
structlog>=23.1.0

# Совместимые версии для Pydantic v2 и FastAPI
pydantic-settings>=2.0.3
pydantic>=2.0.0
fastapi>=0.100.0  # Версия, поддерживающая Pydantic v2
gitpython>=3.1.40
python-dotenv>=1.0.0
fastapi>=0.95.0
Jinja2>=3.1.2